:seeding: improve cli tests
