+++ pkg/cli/cli_test.go
@@ -21,10 +21,89 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 
+	internalconfig "sigs.k8s.io/kubebuilder/internal/config"
 	"sigs.k8s.io/kubebuilder/pkg/model/config"
 	"sigs.k8s.io/kubebuilder/pkg/plugin"
 )
+
+// Test plugin types and constructors.
+type mockPlugin struct {
+	name            string
+	version         plugin.Version
+	projectVersions []string
+}
+
+func (p mockPlugin) Name() string                       { return p.name }
+func (p mockPlugin) Version() plugin.Version            { return p.version }
+func (p mockPlugin) SupportedProjectVersions() []string { return p.projectVersions }
+
+func (mockPlugin) UpdateContext(*plugin.Context) {}
+func (mockPlugin) BindFlags(*pflag.FlagSet)      {}
+func (mockPlugin) InjectConfig(*config.Config)   {}
+func (mockPlugin) Run() error                    { return nil }
+
+func makeBasePlugin(name, version string, projVers ...string) plugin.Base {
+	v, err := plugin.ParseVersion(version)
+	if err != nil {
+		panic(err)
+	}
+	return mockPlugin{name, v, projVers}
+}
+
+func makePluginsForKeys(keys ...string) (plugins []plugin.Base) {
+	for _, key := range keys {
+		n, v := plugin.SplitKey(key)
+		plugins = append(plugins, makeBasePlugin(n, v, internalconfig.DefaultVersion))
+	}
+	return
+}
+
+type mockAllPlugin struct {
+	mockPlugin
+	mockInitPlugin
+	mockCreateAPIPlugin
+	mockCreateWebhookPlugin
+}
+
+type mockInitPlugin struct{ mockPlugin }
+type mockCreateAPIPlugin struct{ mockPlugin }
+type mockCreateWebhookPlugin struct{ mockPlugin }
+
+// GetInitPlugin will return the plugin which is responsible for initialized the project
+func (p mockInitPlugin) GetInitPlugin() plugin.Init { return p }
+
+// GetCreateAPIPlugin will return the plugin which is responsible for scaffolding APIs for the project
+func (p mockCreateAPIPlugin) GetCreateAPIPlugin() plugin.CreateAPI { return p }
+
+// GetCreateWebhookPlugin will return the plugin which is responsible for scaffolding webhooks for the project
+func (p mockCreateWebhookPlugin) GetCreateWebhookPlugin() plugin.CreateWebhook { return p }
+
+func makeAllPlugin(name, version string, projectVersions ...string) plugin.Base {
+	p := makeBasePlugin(name, version, projectVersions...).(mockPlugin)
+	return mockAllPlugin{
+		p,
+		mockInitPlugin{p},
+		mockCreateAPIPlugin{p},
+		mockCreateWebhookPlugin{p},
+	}
+}
+
+func makeSetByProjVer(ps ...plugin.Base) map[string][]plugin.Base {
+	set := make(map[string][]plugin.Base)
+	for _, p := range ps {
+		for _, version := range p.SupportedProjectVersions() {
+			set[version] = append(set[version], p)
+		}
+	}
+	return set
+}
+
+func setPluginsFlag(key string) {
+	os.Args = append(os.Args, "init", "--"+pluginsFlag, key)
+}
 
 var _ = Describe("CLI", func() {
 
@@ -152,10 +231,19 @@ var _ = Describe("CLI", func() {
 			})
 		})
 
+		Context("with extra commands set", func() {
+			It("should work successfully with extra commands", func() {
+				setPluginsFlag("go.test.com/v2")
+				commandTest := &cobra.Command{
+					Use: "example",
+				}
+				c, err = New(WithDefaultPlugins(pluginAV1), WithPlugins(allPlugins...), WithExtraCommands(commandTest))
+				Expect(err).NotTo(HaveOccurred())
+				Expect(c).NotTo(BeNil())
+				Expect(c.(*cli).extraCommands[0]).NotTo(BeNil())
+				Expect(c.(*cli).extraCommands[0].Use).To(Equal(commandTest.Use))
+			})
+		})
 	})
 
 })
-
-func setPluginsFlag(key string) {
-	os.Args = append(os.Args, "init", "--"+pluginsFlag, key)
-}
